requirements.txt
torch==2.3.1
torchvision==0.18.1
accelerate==1.6.0
diffusers==0.32.2
transformers==4.49.0
gradio>=4.0.0
pillow>=9.0.0
numpy>=1.21.0